+++ httpclient/fixtures/testdata/some.json
@@ -1,0 +1,3 @@
+{
+    "some": "data"
+}
